% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{glpm_find_x1}
\alias{glpm_find_x1}
\title{Finds the time-of-flight focal point of the mirror.}
\usage{
glpm_find_x1(L, V)
}
\arguments{
\item{L}{Vector of electrode lengths normalized with the height H of the 
mirror electrodes ("length"/\code{H}).}

\item{V}{Vector of voltages normalized with the mean energy ("voltage"/K_0)}
}
\value{
focal distance, measured from the back plane, normalized with H.
}
\description{
\code{glpm_find_x1} finds the time-of-flight focal point of the mirror.
}
\keyword{internal}
