% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{glpm_integrand}
\alias{glpm_integrand}
\title{Integrand for tof period calculation.}
\usage{
glpm_integrand(x, E, L, V)
}
\arguments{
\item{x}{Axial distance from the end electrode, normalized to the height H of
the mirror electrodes.}

\item{E}{Potential energy at the turning point (normalized with the mean 
energy (K/K_0).}

\item{L}{Vector of electrode lengths normalized with H ("length"/\code{H}).}

\item{V}{Vector of voltages normalized with the mean energy ("voltage"/K_0)}
}
\value{
Axial potential at \code{x}.
}
\description{
\code{glpm_integrand} is the integrand for the time-of-flight period calculation.
}
\keyword{internal}
