% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{glpm_potential}
\alias{glpm_potential}
\title{Calculates the axial potential of gridless planar mirrors.}
\usage{
glpm_potential(x, L, V)
}
\arguments{
\item{x}{Axial distance from the end electrode. x is normalized to the height
H of the mirror electrodes (same height for all electrodes).}

\item{L}{Vector of electrode lengths normalized with H ("length"/\code{H}).}

\item{V}{Vector of voltages normalized with the mean energy ("voltage"/K_0).}
}
\value{
Axial potential at \code{x}.
}
\description{
\code{glpm_potential} calculates the axial potential of gridless planar mirrors.
}
\references{
Yavor, M.I. et al. (2018), High performance gridless ion mirrors 
for multi-reflection time-of-flight and electrostatic trap mass analyzers, 
\emph{International Journal of Mass Spectrometry}, \strong{426},
1-11, doi:10.1016/j.ijms.2018.01.009.
}
\keyword{internal}
