% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{glpm_potential_inv}
\alias{glpm_potential_inv}
\title{Calculates the inverse of the axial potential function.}
\usage{
glpm_potential_inv(y, L, V)
}
\arguments{
\item{y}{Axial potential.}

\item{L}{Vector of electrode lengths normalized with the height H of the 
mirror electrodes ("length"/\code{H}).}

\item{V}{Vector of voltages normalized with the mean energy ("voltage"/K_0)}
}
\value{
Position x (in units of H) where the axial potential = y.
}
\description{
\code{glpm_potential_inv} calculates the position x given the axial potential (inverse
of axial potential function).
}
\details{
The inverse of the axial potential function is used to calculate the turning 
point inside the mirror for a given kinetic energy.
}
\keyword{internal}
