% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imo-package.R
\docType{package}
\name{imo-package}
\alias{imo}
\alias{imo-package}
\title{imo: Ion Mirror Optimization}
\description{
The R package provides functions to optimize the geometry and voltages of planar or axially symmetric ion mirrors, where the potential can be calculated analytically. Design of expriments (DoE) and response surface methodology (RSM) is used for efficient optimization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pasturm/imo}
  \item Report bugs at \url{https://github.com/pasturm/imo/issues}
}

}
\author{
\strong{Maintainer}: Patrick Sturm \email{sturm@tofwerk.com}

Other contributors:
\itemize{
  \item TOFWERK [copyright holder, funder]
}

}
\keyword{internal}
