% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imo-package.R
\docType{package}
\name{imo}
\alias{imo}
\title{Ion Mirror Optimization.}
\description{
The R package provides functions to optimize the geometry and voltages of 
planar or axially symmetric ion mirrors, where the potential can be 
calculated analytically. Design of expriments (DoE) and response surface 
methodology (RSM) is used for efficient optimization.
}
\author{
Patrick Sturm <sturm@tofwerk.com>
}
