% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{pim_find_x1}
\alias{pim_find_x1}
\title{Finds the time-of-flight focal point of the mirror.}
\usage{
pim_find_x1(z1, z2, L, V1, V2)
}
\arguments{
\item{z1}{Distance where electrode 1 starts (normalized with H).}

\item{z2}{Distance where electrode 2 starts (normalized with H).}

\item{L}{Total length of the mirror, i.e. disance where end cap is (normalized with H).}

\item{V1}{Voltage of electrode 1.}

\item{V2}{Voltage of electrode 2.

All distances are normalized with the height of the mirror electrodes H 
(which is the same for all electrodes).}
}
\value{
time-of-flight.
}
\description{
\code{pim_find_x1} finds the time-of-flight focal point of the mirror.
}
\keyword{internal}
