% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{pim_integrand}
\alias{pim_integrand}
\title{Integrand for tof period calculation.}
\usage{
pim_integrand(x, E, z1, z2, L, V1, V2)
}
\arguments{
\item{x}{Axial distance from the entrance grid (normalized with H).}

\item{E}{Potential energy at the turning point (normalized with H).}

\item{z1}{Distance where electrode 1 starts (normalized with H).}

\item{z2}{Distance where electrode 2 starts (normalized with H).}

\item{L}{Total length of the mirror, i.e. disance where end cap is (normalized with H).}

\item{V1}{Voltage of electrode 1.}

\item{V2}{Voltage of electrode 2.

All distances are normalized with the height of the mirror electrodes H 
(which is the same for all electrodes).}
}
\value{
Axial potential at \code{x}.
}
\description{
\code{pim_integrand} is the integrand for the time-of-flight period calculation.
}
\keyword{internal}
