% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{pim_potential}
\alias{pim_potential}
\title{Calculates the axial potential of a planar three-element ion mirror.}
\usage{
pim_potential(x, z1, z2, L, V1, V2)
}
\arguments{
\item{x}{Axial distance from the entrance grid (normalized with H).}

\item{z1}{Distance where electrode 1 starts (normalized with H).}

\item{z2}{Distance where electrode 2 starts (normalized with H).}

\item{L}{Total length of the mirror, i.e. disance where end cap is (normalized with H).}

\item{V1}{Voltage of electrode 1.}

\item{V2}{Voltage of electrode 2.

All distances are normalized with the height of the mirror electrodes H 
(which is the same for all electrodes).}
}
\value{
Axial potential at \code{x}.
}
\description{
\code{pim_potential} calculates the axial potential of a planar 
three-element ion mirror.
}
\references{
Yavor, M.I. et al. (2018), High performance gridless ion mirrors 
for multi-reflection time-of-flight and electrostatic trap mass analyzers, 
\emph{International Journal of Mass Spectrometry}, \strong{426},
1-11, doi:10.1016/j.ijms.2018.01.009.
}
\keyword{internal}
