% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{pim_potential_inv}
\alias{pim_potential_inv}
\title{Calculates the inverse of the axial potential function.}
\usage{
pim_potential_inv(y, z1, z2, L, V1, V2)
}
\arguments{
\item{z1}{Distance where electrode 1 starts (normalized with H).}

\item{z2}{Distance where electrode 2 starts (normalized with H).}

\item{L}{Total length of the mirror, i.e. disance where end cap is (normalized with H).}

\item{V1}{Voltage of electrode 1.}

\item{V2}{Voltage of electrode 2.

All distances are normalized with the height of the mirror electrodes H 
(which is the same for all electrodes).}

\item{x}{Axial distance from the entrance grid (normalized with H).}
}
\value{
Position x where the axial potential = y.
}
\description{
\code{pim_potential_inv} calculates the position x given the axial potential (inverse
of axial potential function).
}
\details{
The inverse of the axial potential function is used to calculate the turning 
point inside the mirror for a given kinetic energy.
}
\keyword{internal}
