% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{pim_tofperiod}
\alias{pim_tofperiod}
\title{Time-of-flight period calculation.}
\usage{
pim_tofperiod(E, z, V)
}
\arguments{
\item{E}{Potential energy at the turning point.}

\item{z}{Vector of electrode distances (normalized with H). \code{z[1]} it
the end of the entrance grid electrode and \code{z[length(z)]} is the total
length of the mirror.}

\item{V}{Vector of electrode voltages. The entrance grid and the first
electrode are assumed to be grounded (\code{V[1]=0}).

All distances are normalized with the height of the mirror electrodes H 
(which is the same for all electrodes).}
}
\value{
time-of-flight.
}
\description{
\code{pim_tofperiod} calculates the time-of-flight period inside the mirror.
}
\details{
This integrates the particle motion from the entrance grid to the turning 
point x0. Because here we are only interested in relative time-of-flight 
deviations all constant factors (e.g. \code{sqrt(m*amu/e/2)}) are omitted 
and the time-of-flight is in arbitrary units.
}
\keyword{internal}
