% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{pim_totaltof}
\alias{pim_totaltof}
\title{Time-of-flight period calculation.}
\usage{
pim_totaltof(E, z, V, x1, d5 = 0, u5 = 0)
}
\arguments{
\item{E}{Potential energy at the turning point.}

\item{z}{Vector of electrode distances (normalized with H). \code{z[1]} it
the end of the entrance grid electrode and \code{z[length(z)]} is the total
length of the mirror.}

\item{V}{Vector of normalized electrode potentials. The entrance grid and the 
first electrode are assumed to be grounded (\code{V[1]=0}).}

\item{x1}{Focal distance of mirror (normalized with H), measured from the 
entrance grid).}

\item{d5}{Distance of first linear stage of mirror (normalized with H).}

\item{u5}{Normalized potential in first linear stage of mirror.

All distances are normalized to the height of the mirror electrodes H 
(which is the same for all electrodes). The potentials are normalized to
the mean kinetic energy ("voltage"/K_0) of the (singly-charged) ions with
\code{u5=0}. With \code{u5!=0}, the normalized kinetic energy is \code{1+u5}.}
}
\value{
time-of-flight.
}
\description{
\code{pim_totaltof} calculates the total time-of-flight including field-free
space and linear stage of the mirror.
}
\details{
Because here we are only interested in relative time-of-flight 
deviations all constant factors (e.g. \code{sqrt(m*amu/e/2)}) are omitted 
and the time-of-flight is in arbitrary units.
}
\keyword{internal}
