% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imo.R
\name{run_imo}
\alias{run_imo}
\title{Runs ion mirror optimization.}
\usage{
run_imo(
  imo_config,
  type = c("GLPM", "ZEIM", "PIM"),
  resume = FALSE,
  write = TRUE,
  plot = TRUE
)
}
\arguments{
\item{imo_config}{path and name of imo_config file (.toml)}

\item{type}{Ion mirror type. Currently supported are \code{"GLPM"}: gridless
planar mirrors, \code{"ZEIM"}: cylindrical Zhang-Enke three-element ion 
mirrors, \code{"PIM"}: planar ion mirrors.}

\item{resume}{if \code{FALSE} (default) starting values are taken from the
imo_config file, if \code{TRUE} starting values are taken from the last
bestpoint_run.txt (see 'Details').}

\item{write}{Write output files (\code{TRUE} (default) or \code{FALSE}).}

\item{plot}{Plot results (\code{TRUE} (default) or \code{FALSE}).}
}
\value{
A data.frame containing the optimized best points from the last run.
}
\description{
\code{run_imo} runs ion mirror optimization.
}
\details{
The experiment needs to be configured in a imo_config file (TOML file format).
See \code{system.file("GLPMtuneR_config.toml", package = "imo")} for 
a template.

If you want to start from a previous best point, set \code{resume = TRUE}.
If there is a bestpoint_run.txt file in the imo_results directory, the starting 
values will be taken from there. Otherwise the latest bestpoint_run.txt from
all subfolders of imo_results will be taken.
}
\examples{
\dontrun{
imo_config = system.file("ZEIMimo_config.toml", package = "imo")
bestpoint = run_imo(imo_config, type = "ZEIM")
}

}
