% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imo.R
\name{run_PIMo}
\alias{run_PIMo}
\title{Runs planar ion mirror (PIM) optimization.}
\usage{
run_PIMo(imo_config, resume = FALSE)
}
\arguments{
\item{imo_config}{path and name of imo_config file (.toml)}

\item{resume}{if \code{FALSE} (default) starting values are taken from the
tuneR_config file, if \code{TRUE} starting values are taken from the last
bestpoint_run.txt (see 'Details').}
}
\description{
\code{run_PIMo} runs gridless planar mirror optimization.
}
\details{
The experiment needs to be configured in a imo_config file (TOML file format).
See \code{system.file("GLPMtuneR_config.toml", package = "imo")} for 
a template.

If you want to start from a previous best point, set \code{resume = TRUE}.
If there is a bestpoint_run.txt file in the imo_results directory, the starting 
values will be taken from there. Otherwise the latest bestpoint_run.txt from
all subfolders of imo_results will be taken.
}
\examples{
\dontrun{
imo_config = system.file("PIMimo_config.toml", package = "imo")
run_PIMo(imo_config)
}

}
