% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{snlm_potential}
\alias{snlm_potential}
\title{Calculates the potential of a Shimadzu non-linear mirror.}
\usage{
snlm_potential(d1, d3, d4, d5, d6, shift, drift, pulse)
}
\arguments{
\item{d1}{Distance of first stage of extractor, from push to pull (m).}

\item{d3}{Distance of second stage of extractor, from pull to drift (m).}

\item{d4}{Distance of of drift space, from extractor to reflector and from
reflector to detector (m).}

\item{d5}{Distance of first stage of reflector (m).}

\item{d6}{Distance of second stage of reflector (m).}

\item{shift}{Reference energy shift inside extractor (m).}

\item{drift}{Drift tube voltage (V).}

\item{pulse}{Extraction pulse voltage (V).}
}
\value{
A data.frame with the corrected total distance, the corresponding 
potential and the distance correction in the second stage of the mirror.
}
\description{
\code{snlm_potential} calculates the potential of a Shimadzu non-linear 
mirror, which is a linear two-stage ion mirror with a non-linear correction
potential in the second stage.
}
\references{
U.S. Patent No. 8,772,708 B2 (filed Dec. 20, 2011).
}
\keyword{internal}
