% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{zeim_find_x1}
\alias{zeim_find_x1}
\title{Finds the time-of-flight focal point of the mirror.}
\usage{
zeim_find_x1(z1, z2, L, V1, V2, R)
}
\arguments{
\item{z1}{Distance where electrode 1 starts.}

\item{z2}{Distance where electrode 2 starts.}

\item{L}{Total length of the mirror, i.e. disance where end cap is.}

\item{V1}{Voltage of electrode 1.}

\item{V2}{Voltage of electrode 2.}

\item{R}{Radius of the cylinder.}
}
\value{
focal point (measured from the entrance grid).
}
\description{
\code{zeim_find_x1} finds the time-of-flight focal point of the mirror.
}
\keyword{internal}
