% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{zeim_integrand}
\alias{zeim_integrand}
\title{Integrand for tof period calculation.}
\usage{
zeim_integrand(x, E, z1, z2, L, V1, V2, R)
}
\arguments{
\item{x}{Axial distance from the entrance grid.}

\item{E}{Potential energy at the turning point.}

\item{z1}{Distance where electrode 1 starts.}

\item{z2}{Distance where electrode 2 starts.}

\item{L}{Total length of the mirror, i.e. disance where end cap is.}

\item{V1}{Voltage of electrode 1.}

\item{V2}{Voltage of electrode 2.}

\item{R}{Radius of the cylinder.}
}
\value{
Axial potential at \code{x}.
}
\description{
\code{zeim_integrand} is the integrand for the time-of-flight period calculation.
}
\keyword{internal}
