% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{zeim_potential}
\alias{zeim_potential}
\title{Calculates the axial potential of cylindrical Zhang-Enke ion mirrors.}
\usage{
zeim_potential(x, z1, z2, L, V1, V2, R)
}
\arguments{
\item{x}{Axial distance from the entrance grid.}

\item{z1}{Distance where electrode 1 starts.}

\item{z2}{Distance where electrode 2 starts.}

\item{L}{Total length of the mirror, i.e. disance where end cap is.}

\item{V1}{Voltage of electrode 1.}

\item{V2}{Voltage of electrode 2.}

\item{R}{Radius of the cylinder.}
}
\value{
Axial potential at \code{x}.
}
\description{
\code{zeim_potential} calculates the axial potential of cylindrical 
three-element Zhang-Enke ion mirrors.
}
\references{
Yefchak, G.E. and Flory, C.A. (2002), Improved method for 
designing a cylindrical Zhang-Enke ion mirror, \emph{International Journal 
of Mass Spectrometry}, \strong{214}, 89-94, doi:10.1016/S1387-3806(01)00564-4.
}
\keyword{internal}
