% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{zeim_potential_inv}
\alias{zeim_potential_inv}
\title{Calculates the inverse of the axial potential function.}
\usage{
zeim_potential_inv(y, z1, z2, L, V1, V2, R)
}
\arguments{
\item{z1}{Distance where electrode 1 starts.}

\item{z2}{Distance where electrode 2 starts.}

\item{L}{Total length of the mirror, i.e. disance where end cap is.}

\item{V1}{Voltage of electrode 1.}

\item{V2}{Voltage of electrode 2.}

\item{R}{Radius of the cylinder.}

\item{x}{Axial distance from the entrance grid.}
}
\value{
Position x where the axial potential = y.
}
\description{
\code{zeim_potential_inv} calculates the position x given the axial potential (inverse
of axial potential function).
}
\details{
The inverse of the axial potential function is used to calculate the turning 
point inside the mirror for a given kinetic energy.
}
\keyword{internal}
