% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{zeim_tofperiod}
\alias{zeim_tofperiod}
\title{Time-of-flight period calculation.}
\usage{
zeim_tofperiod(E, z1, z2, L, V1, V2, R)
}
\arguments{
\item{E}{Potential energy at the turning point.}

\item{z1}{Distance where electrode 1 starts.}

\item{z2}{Distance where electrode 2 starts.}

\item{L}{Total length of the mirror, i.e. disance where end cap is.}

\item{V1}{Voltage of electrode 1.}

\item{V2}{Voltage of electrode 2.}

\item{R}{Radius of the cylinder.}
}
\value{
time-of-flight.
}
\description{
\code{zeim_tofperiod} calculates the time-of-flight period inside the mirror.
}
\details{
This integrates the particle motion from the entrance grid to the turning 
point x0. Because here we are only interested in relative time-of-flight 
deviations all constant factors (e.g. \code{sqrt(m*amu/e/2)}) are omitted 
and the time-of-flight is in arbitrary units.
}
\keyword{internal}
