% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{glpm_find_x1}
\alias{glpm_find_x1}
\title{Finds the time-of-flight focal point of the mirror.}
\usage{
glpm_find_x1(L, V, H)
}
\arguments{
\item{L}{Vector of electrode lengths normalized with H ("lenght"/\code{H}).}

\item{V}{Vector of voltages normalized with the mean energy ("voltage"/K_0)}

\item{H}{Height of the mirror electrodes (the same for all electrodes).}
}
\value{
time-of-flight.
}
\description{
\code{glpm_find_x1} finds the time-of-flight focal point of the mirror.
}
\keyword{internal}
