% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{glpm_tofperiod}
\alias{glpm_tofperiod}
\title{Time-of-flight period calculation.}
\usage{
glpm_tofperiod(E, x1, L, V)
}
\arguments{
\item{E}{Vector of potential energies at the turning point (normalized with
the mean energy (K/K_0).}

\item{x1}{End distance of the particle, typically the time-of-flight focal 
point of the mirror (in units of H).}

\item{L}{Vector of electrode lengths normalized with the height H of the 
mirror electrodes ("length"/\code{H}).}

\item{V}{Vector of voltages normalized with the mean energy ("voltage"/K_0)}
}
\value{
time-of-flight.
}
\description{
\code{glpm_tofperiod} calculates the time-of-flight period inside the mirror.
}
\details{
This integrates the particle motion from the turning point x0 to the distance
x1. Because here we are only interested in relative time-of-flight deviations
all constant factors (e.g. \code{sqrt(m*amu/e/2)}) are omitted and the 
time-of-flight is in arbitrary units.
}
\keyword{internal}
