% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{pim_find_x1}
\alias{pim_find_x1}
\title{Finds the time-of-flight focal point of the mirror.}
\usage{
pim_find_x1(z, V)
}
\arguments{
\item{z}{Vector of electrode distances (normalized with H). \code{z[1]} it
the end of the entrance grid electrode and \code{z[length(z)]} is the total
length of the mirror.}

\item{V}{Vector of electrode voltages. The entrance grid and the first
electrode are assumed to be grounded (\code{V[1]=0}).

All distances are normalized with the height of the mirror electrodes H 
(which is the same for all electrodes).}
}
\value{
time-of-flight.
}
\description{
\code{pim_find_x1} finds the time-of-flight focal point of the mirror.
}
\keyword{internal}
