% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{pim_integrand}
\alias{pim_integrand}
\title{Integrand for tof period calculation.}
\usage{
pim_integrand(x, E, z, V)
}
\arguments{
\item{x}{Axial distance from the entrance grid (normalized with H).}

\item{E}{Potential energy at the turning point (normalized with H).}

\item{z}{Vector of electrode distances (normalized with H). \code{z[1]} it
the end of the entrance grid electrode and \code{z[length(z)]} is the total
length of the mirror.}

\item{V}{Vector of electrode voltages. The entrance grid and the first
electrode are assumed to be grounded (\code{V[1]=0}).

All distances are normalized with the height of the mirror electrodes H 
(which is the same for all electrodes).}
}
\value{
Axial potential at \code{x}.
}
\description{
\code{pim_integrand} is the integrand for the time-of-flight period calculation.
}
\keyword{internal}
