% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{pim_potential}
\alias{pim_potential}
\title{Calculates the axial potential of a planar multi-element ion mirror.}
\usage{
pim_potential(x, z, V)
}
\arguments{
\item{x}{Axial distance from the entrance grid (normalized with H).}

\item{z}{Vector of electrode distances (normalized with H). \code{z[1]} is
the end of the entrance grid electrode and \code{z[length(z)]} is the total
length of the mirror.}

\item{V}{Vector of electrode voltages. The entrance grid and the first
electrode are assumed to be grounded (\code{V[1]=0}).

All distances are normalized with the height of the mirror electrodes H 
(which is the same for all electrodes).}
}
\value{
Axial potential at \code{x}.
}
\description{
\code{pim_potential} calculates the axial potential of a planar 
multi-element ion mirror with entrance grid.
}
\references{
Yavor, M.I. et al. (2018), High performance gridless ion mirrors 
for multi-reflection time-of-flight and electrostatic trap mass analyzers, 
\emph{International Journal of Mass Spectrometry}, \strong{426},
1-11, doi:10.1016/j.ijms.2018.01.009.
}
\keyword{internal}
