% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{pim_potential_inv}
\alias{pim_potential_inv}
\title{Calculates the inverse of the axial potential function.}
\usage{
pim_potential_inv(y, z, V)
}
\arguments{
\item{z}{Vector of electrode distances (normalized with H). \code{z[1]} it
the end of the entrance grid electrode and \code{z[length(z)]} is the total
length of the mirror.}

\item{V}{Vector of normalized electrode potentials. The entrance grid and the 
first electrode are assumed to be grounded (\code{V[1]=0}).

All distances are normalized to the height of the mirror electrodes H 
(which is the same for all electrodes) and the potentials are normalized to
the mean kinetic energy ("voltage"/K_0) of the (singly-charged) ions.}

\item{x}{Axial distance from the entrance grid (normalized with H).}
}
\value{
Position x where the axial potential = y.
}
\description{
\code{pim_potential_inv} calculates the position x given the axial potential (inverse
of axial potential function).
}
\details{
The inverse of the axial potential function is used to calculate the turning 
point inside the mirror for a given kinetic energy.
}
\keyword{internal}
