% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{pim_tofperiod}
\alias{pim_tofperiod}
\title{Time-of-flight period calculation.}
\usage{
pim_tofperiod(E, z1, z2, L, V1, V2)
}
\arguments{
\item{E}{Potential energy at the turning point.}

\item{z1}{Distance where electrode 1 starts (normalized with H).}

\item{z2}{Distance where electrode 2 starts (normalized with H).}

\item{L}{Total length of the mirror, i.e. disance where end cap is (normalized with H).}

\item{V1}{Voltage of electrode 1.}

\item{V2}{Voltage of electrode 2.

All distances are normalized with the height of the mirror electrodes H 
(which is the same for all electrodes).}
}
\value{
time-of-flight.
}
\description{
\code{pim_tofperiod} calculates the time-of-flight period inside the mirror.
}
\details{
This integrates the particle motion from the entrance grid to the turning 
point x0. Because here we are only interested in relative time-of-flight 
deviations all constant factors (e.g. \code{2*sqrt(2*m*amu/e)}) are omitted 
and the time-of-flight is in arbitrary units.
}
\keyword{internal}
