% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{pim_totaltof}
\alias{pim_totaltof}
\title{Time-of-flight period calculation.}
\usage{
pim_totaltof(E, z, V, x1, d5 = 0, u5 = 0)
}
\arguments{
\item{E}{Potential energy at the turning point.}

\item{z}{Vector of electrode distances (normalized with H). \code{z[1]} it
the end of the entrance grid electrode and \code{z[length(z)]} is the total
length of the mirror.}

\item{V}{Vector of electrode voltages. The entrance grid and the first
electrode are assumed to be grounded (\code{V[1]=0}).}

\item{x1}{Focal distance of mirror (normalized with H).}

\item{d5}{Distance of first linear stage of mirror (normalized with H).}

\item{u5}{Potential in first linear stage of mirror.

All distances are normalized with the height of the mirror electrodes H 
(which is the same for all electrodes).}
}
\value{
time-of-flight.
}
\description{
\code{pim_totaltof} calculates the total time-of-flight including field-free
space and linear stage of the mirror.
}
\details{
Because here we are only interested in relative time-of-flight 
deviations all constant factors (e.g. \code{2*sqrt(2*m*amu/e)}) are omitted 
and the time-of-flight is in arbitrary units.
}
\keyword{internal}
